refactor: 프로젝트를 후원한 기업 랭킹을 조회하는 기능 기업에서 프로젝트로 이동
