+++ src/main/java/boosters/fundboost/project/repository/CustomProjectRepository.java
@@ -1,0 +1,10 @@
+package boosters.fundboost.project.repository;
+
+import com.querydsl.core.Tuple;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+
+public interface CustomProjectRepository {
+    Page<Tuple> findBoostedCompanies(long projectId, Pageable pageable);
+}
