+++ src/main/java/boosters/fundboost/project/service/ProjectService.java
@@ -1,5 +1,7 @@
 package boosters.fundboost.project.service;
 
+import boosters.fundboost.company.dto.request.CompanyRankingPreviewRequest;
+import boosters.fundboost.company.dto.response.CompanyRankingPreviewResponse;
 import boosters.fundboost.project.domain.enums.ProjectCategory;
 import boosters.fundboost.project.domain.enums.Region;
 import boosters.fundboost.project.dto.request.ProjectBasicInfoRequest;
@@ -13,9 +15,9 @@ import java.util.List;
 
 public interface ProjectService {
     void registerBasicInfo(ProjectBasicInfoRequest request, MultipartFile image);
-  
+
     void updateProject(Long projectId, ProjectBasicInfoRequest request, MultipartFile image);
-  
+
     void deleteProject(Long projectId);
 
     List<NewProjectResponse> getNewProjects();
@@ -35,4 +37,6 @@ public interface ProjectService {
     ProjectDetailResponse getProjectDetail(Long projectId);
 
     long getProjectCount(String getType);
+
+    Page<CompanyRankingPreviewResponse> getBoostedCompanyRanking(CompanyRankingPreviewRequest request);
 }
